// frontend/src/api/generated/model/extendedDetectionConfigSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * astroimage
 * OpenAPI spec version: 0.1.0
 */

export interface ExtendedDetectionConfigSchema {
  /** @minimum 0.5 */
  sigma?: number;
  /** @minimum 0.5 */
  smooth_sigma?: number;
  /** @minimum 1 */
  min_area?: number;
  /** @minimum 0 */
  max_area?: number;
  /** @minimum 1 */
  bin_factor?: number;
  /** @minimum 0 */
  closing_iterations?: number;
  /** @minimum 0 */
  opening_iterations?: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  min_score?: number;
  /** @minimum 0 */
  max_sources?: number;
}