// frontend/src/api/generated/model/pointDetectionConfigSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * astroimage
 * OpenAPI spec version: 0.1.0
 */

export interface PointDetectionConfigSchema {
  /** @minimum 0.5 */
  fwhm?: number;
  /** @minimum 1 */
  sigma?: number;
  /** @minimum 0 */
  min_snr?: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  min_score?: number;
  /** @minimum 0 */
  min_distance?: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  visual_weight?: number;
  /** @minimum 1 */
  visual_area_radius?: number;
  /** @minimum 0 */
  visual_area_sigma?: number;
  /** @minimum 0 */
  max_sources?: number;
}