// frontend/src/api/generated/model/sourcePresetResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * astroimage
 * OpenAPI spec version: 0.1.0
 */
import type { ExtendedDetectionConfigSchema } from './extendedDetectionConfigSchema';
import type { PointDetectionConfigSchema } from './pointDetectionConfigSchema';
import type { PresetChangeSchema } from './presetChangeSchema';
import type { PresetEvidenceSchema } from './presetEvidenceSchema';

export interface SourcePresetResponse {
  point_config: PointDetectionConfigSchema;
  extended_config: ExtendedDetectionConfigSchema;
  baseline_point_config: PointDetectionConfigSchema;
  baseline_extended_config: ExtendedDetectionConfigSchema;
  evidence?: PresetEvidenceSchema;
  changes?: PresetChangeSchema[];
}